third_party/WebKit/LayoutTests/fast/viewport/viewport-warnings-2-expected.txt

CONSOLE ERROR: line 3: The key "wwidth" is not recognized and ignored.
ALERT: viewport size 980x1078 scale 0.326531 with limits [0.326531, 5] and userScalable true
